+++ app/src/main/java/com/hengye/share/module/sina/WBTopicIds.java
@@ -18,7 +18,7 @@ public class WBTopicIds {
     private boolean hasvisible;
     private long previous_cursor;
     private long next_cursor;
-    private int total_number;
+    private long total_number;
     private int interval;
     private int uve_blank;
     private List<String> statuses;
@@ -52,7 +52,7 @@ public class WBTopicIds {
         this.next_cursor = next_cursor;
     }
 
-    public void setTotal_number(int total_number) {
+    public void setTotal_number(long total_number) {
         this.total_number = total_number;
     }
 
@@ -88,7 +88,7 @@ public class WBTopicIds {
         return next_cursor;
     }
 
-    public int getTotal_number() {
+    public long getTotal_number() {
         return total_number;
     }
